'fix:bugs'

+++ app/Http/Controllers/ServiceController.php
@@ -42,10 +42,6 @@ class ServiceController extends Controller
 
     public function store(ServiceRequest $request)
     {
-        try {
-            if (auth()->user()->role != 1) {
-                abort(403, 'Unauthorized Action');
-            }
 
             $fileName = time() . '_' . $request->logo->getClientOriginalName();
             $filePath = $request->file('logo')->storeAs('uploads', $fileName, 'public');
@@ -58,31 +54,33 @@ class ServiceController extends Controller
         }
     }
 
-    public function edit(Service $service)
-    {
-        try {
-            return view('site.services.edit', ['service' => $service]);
-        } catch (Exception $e) {
-            throw new Exception($e->getMessage());
+
+    public function edit(Service $service) {
+        if( auth()->user()->role != 1) {
+            abort(403, 'Unauthorized Action');
         }
+        return view('site.services.edit',['service' => $service]);
     }
 
-    public function update(ServiceRequest $request, Service $service)
-    {
-        try {
-            $fileName = time() . '_' . $request->logo->getClientOriginalName();
-            $filePath = $request->file('logo')->storeAs('uploads', $fileName, 'public');
+    public function update(ServiceRequest $request, Service $service) {
+        $fileName = time() . '_' . $request->logo->getClientOriginalName();
+        $filePath = $request->file('logo')->storeAs('uploads', $fileName, 'public');
 
-            $service->update(array_merge($request->validated(), ['logo' => $filePath]));
-            return redirect('/')->with('message', 'Service Updated successfully');
+        $service->update(array_merge($request->validated(), ['logo' => $filePath]));
+        return redirect('/')->with('message', 'Service Updated successfully');
+
         } catch (Exception $e) {
             throw new Exception($e->getMessage());
         }
+
     }
 
     public function destroy(Service $service)
     {
         try {
+        if( auth()->user()->role != 1) {
+            abort(403, 'Unauthorized Action');
+        }
             $service->delete();
             return redirect('/')->with('message', 'Service deleted successfully');
         } catch (Exception $e) {
